requirements.txt
Flask-WTF==1.2.2
WTForms==3.2.1
requests==2.32.3
pandas==2.2.3
seaborn==0.13.2
plotly==6.0.1
matplotlib==3.10.1
python-dotenv==1.1.0
Flask==3.1.0
Flask-Bootstrap==2.4.2